merge

+++ Assets/ScriptableObjects/Weapon/WeaponData.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public abstract class WeaponData : ResourceObject
 {
-    public float damage = 1;
+    public int damage = 1;
     public float attackRate = 1;
     public string audioClipName = "ShootDefault";
 }
